FIx footer covering content
The entire ".App" div is now a flex container

Appears to be working properly on all pages, except:
1. On Blogs page, it appears that the footer is covering the FAB.
But the FAB is still on top of the footer, it's its color that makes it
appear so.

The fix is a significant change. The App div is changed to MUI Box,
and the Header, Footer and  Switch are Box items inside it.
And the only way I found to make the footer stay at the bottom was to
add a minHeight to the main content (Switch)